requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
flake8==3.8.3
future==0.18.2
idna==2.10
line-bot-sdk==1.17.0
mccabe==0.6.1
mysqlclient==2.0.1
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
